(ft)test maid-request [finishes:#170973708]
