+++ .babelrc
@@ -1,3 +1,15 @@
 {
-    "presets": ["@babel/preset-env"]
+    "presets": ["@babel/preset-env"],
+    "plugins":[
+      [
+          "@babel/plugin-transform-runtime",
+      {
+          "absoluteRuntime": true,
+          "corejs": false,
+          "helpers": true,
+          "regenerator": true,
+          "useESModules": false
+      }
+      ]
+  ]
   }
